[MobileHC-4]: Fix: fix register UI design.
