+++ lib/presentation/routes/app_router.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_health_care/presentation/routes/route_paths.dart';
 import 'package:go_router/go_router.dart';
 
-import '../auth/src.dart';
+import '../auth/ui/src.dart';
 
 final GoRouter appRouter = GoRouter(
   initialLocation: RoutePaths.register, // RoutePaths.splash,
